got compiling on upuntu 17

+++ server_files/srcs/cli_lst.c
@@ -62,7 +62,7 @@ void	move_cli(t_client *client, t_direction direction)
 {
 	delete_cli(&MAP(client->pos.x, client->pos.y), client);
 	if (direction == north)
-		client->pos.y = (client->pos.y == 0) ? G_WIDTH - 1 : \
+		client->pos.y = (client->pos.y == 0) ? (int)G_WIDTH - 1 : \
 			client->pos.y - 1;
 	else if (direction == east)
 		client->pos.x = ((int)G_WIDTH - 1 == client->pos.x) ? \
@@ -71,7 +71,7 @@ void	move_cli(t_client *client, t_direction direction)
 		client->pos.y = ((int)G_WIDTH - 1 == client->pos.y) ? \
 			0 : client->pos.y + 1;
 	else if (direction == west)
-		client->pos.x = (client->pos.x == 0) ? G_WIDTH - 1 : \
+		client->pos.x = (client->pos.x == 0) ? (int)G_WIDTH - 1 : \
 			client->pos.x - 1;
 	new_cli(&MAP(client->pos.x, client->pos.y));
 }
